llmnrd: Use MAXHOSTNAMELEN for hostname buffer and zero-terminate string
Use MAXHOSTNAMELEN defined in sys/param.h to determine the size of the
hostname buffer to allocate. Also make sure the string is properly
zero-terminated in all cases, as in some cases it might be undefined
whether a truncated string is properly terminated.

Note: Also remove a TODO comment, as watching for hostname changes at
runtime is outside the scope of llmnrd (also, avahi behaves the same).

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>