llmnrd: merge rtnl interface event loop into main select() loops
Instead of spawming a thread for the sole purpose of watching the rtnl
for (presumably) seldom events, just merge the select() for all sockets
together in one single main loop.

This reduces unnecessary complexity and makes llmnrd no longer require
any locking.

It also allows us to handle signals in a thread-safe manner (as there
aren't any ;) and thus the race condition on exit reported in #20

Closes #20
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>