llmnr-query: Allocate receive buffer large enough
If an LLMNR response exceeds 128 bytes, it will get truncated. This
leads to out-of-bounds read access during parsing and causes garbage
data to be printed.

Fix it by allocating the buffer large enough (according to RFC).
Possible future improvment would be to consider the link MTU size to
save some memory.

Fixes #21
Signed-off-by: Tobias Klauser <tklauser@distanz.ch>